requirements.txt
opencv-contrib-python
tensorflow-gpu